refactor(components/inputs): change textfield props and separate the logic from input
