+++ src/components/inputs/password/index.tsx
@@ -5,8 +5,6 @@ import { VisibilityOutlined, VisibilityOffOutlined } from '@material-ui/icons'
 
 import { TextField, TextFieldProps } from '@/components'
 
-export type PasswordInputProps = TextFieldProps
-
 export const PasswordInput: React.FC<PasswordInputProps> = (props) => {
   const [showPassword, setShowPassword] = useState(false)
 
@@ -23,10 +21,14 @@ export const PasswordInput: React.FC<PasswordInputProps> = (props) => {
   return (
     <TextField
       {...props}
-      inputElementProps={{
-        type: showPassword ? 'text' : 'password'
+      materialInputProps={{
+        inputProps: {
+          type: showPassword ? 'text' : 'password'
+        }
       }}
       icon={icon}
     />
   )
 }
+
+export type PasswordInputProps = TextFieldProps
